move utils, validators to service module
